Move some tests from test_onnx
